Add option to ignore tracks with a given substring.
Add the --ignore option to the setdefaultbylang command. This option
takes a substring and checks in a case-insentive way if it appears
inside the name of each track matching the specified language. The
program ignores the track if the substring is present, even if the
language matches. This is useful to ignore 'forced' subtitle tracks that
appear before regular subtitle tracks.

This option can be used multiple times to ignore multiple substrings.